Refactoring attempt in using hardware PWM
I've been unsatisfied with the non-use of PWM and the lack of UI
responsiveness eversince. Rewired the hardware (switched RESET and STEP)
to be able to use OC1B as hardware output. Frequency generation works,
but is glitchy sometimes when changing the ICR/OCR registers and I
cannot figure why that is. Assuming that there's a race condition
between the timer counting and software update of the registers.